add middleware for checking duplicates
